🐛 Fix healthcheck command in docker-compose.yaml
The '$$' sign was added to the healthcheck command that tests Postgres database readiness to correctly interpret the shell syntax in the Docker configuration file. This will ensure correct interpretation of the $POSTGRES_USER syntax as a shell variable and not as a string literal, improving the health checking of the Postgres service in our docker system.