✨ Adjust pythonpath and file location in docker-compose files
Changes have been made in the 'iii-devops-lite-api' service to set PYTHONPATH in the command, ensuring that the correct path is used when running Celery. This was causing the Celery worker and beat services to fail unexpectedly as they were not able to locate the required files. In addition, the location of the HEAD.env file in the 'docker-compose.iiidevops.yaml' has been adjusted to correctly map the file in the Docker container.